# specs/001-artist-dashboard/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: EVEN Artist Dashboard API
  description: |
    Mock API for the EVEN Artist Dashboard take-home project.

    EVEN is a direct-to-fan platform that lets music artists sell music, videos,
    merch, and experiences directly to fans while retaining more revenue.
  version: 1.0.0
  contact:
    name: EVEN
    url: https://get.even.biz
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /products:
    get:
      summary: Get products
      description: Returns a list of the artist's products with optional filtering and sorting
      operationId: getProducts
      tags:
        - Products
      parameters:
        - name: limit
          in: query
          description: Maximum number of products to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 12
        - name: category
          in: query
          description: Filter by product category
          schema:
            type: string
            enum: [music, video, merch, experience, all]
            default: all
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [newest, bestselling]
            default: newest
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                    description: Total number of products matching filter
                required:
                  - products
                  - total
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    ProductCategory:
      type: string
      enum: [music, video, merch, experience]
      description: EVEN product categories
    Product:
      type: object
      description: A product listed for sale on EVEN
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        category:
          $ref: '#/components/schemas/ProductCategory'
        imageUrl:
          type: string
          format: uri
        listingDate:
          type: string
          format: date-time
        price:
          type: integer
          minimum: 0
          description: Price in cents
        totalEarnings:
          type: integer
          minimum: 0
          description: Artist earnings in cents
        unitsSold:
          type: integer
          minimum: 0
        # Music-specific fields
        releaseType:
          type: string
          enum: [single, ep, album]
          description: Only for music products
        trackCount:
          type: integer
          minimum: 1
          description: Only for music products
        streams:
          type: integer
          minimum: 0
          description: Only for music products
        # Video-specific fields
        duration:
          type: integer
          minimum: 1
          description: Video duration in seconds
        viewCount:
          type: integer
          minimum: 0
          description: Only for video products
        # Merch-specific fields
        variant:
          type: string
          description: Size/color variant for merch
        inventory:
          type: integer
          minimum: 0
          description: Remaining stock for merch
        # Experience-specific fields
        eventDate:
          type: string
          format: date-time
          description: Event date for experiences
        capacity:
          type: integer
          minimum: 1
          description: Max attendees for experiences
        attendees:
          type: integer
          minimum: 0
          description: Current signups for experiences
      required:
        - id
        - title
        - category
        - imageUrl
        - listingDate
        - price
        - totalEarnings
        - unitsSold
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        requestId:
          type: string
          format: uuid
          description: Request correlation ID for debugging
        statusCode:
          type: integer
          minimum: 400
          maximum: 599
      required:
        - error
        - message
        - requestId
        - statusCode
...